Fix the loader not loading assets (see description)
This seems to be an issue with `new Loader()` as the shared loader works
perfectly fine. This means that either everything will need to use the
shared loader and that future versions of loadAssets and loadAsset will
need to check the shared loader for whether the assets are loaded or
not, or someone will need to do some technical reasearch and figure out
what's going on. Until then pixi.js will throw an error every time one
of the methods try to load an existing asset.